update lib structure
