+++ src/lib/src/cli/mod.rs
@@ -1,0 +1,8 @@
+pub mod dialog;
+pub use dialog::*;
+pub mod setup;
+pub use setup::*;
+pub mod uninstall;
+pub use uninstall::*;
+pub mod update_cli;
+pub use update_cli::*;
