+++ src/lib/src/cli/setup.rs
