+++ src/lib/src/cli/uninstall.rs
