+++ src/lib/src/cli/update_cli.rs
@@ -3,8 +3,8 @@ use anyhow::Result;
 use crate::{async_command, check_env, check_repo, setup_work_dir, CARDANO_URL};
 
 pub async fn update_cli() -> Result<()> {
+    log::info!("Updating the CLI");
     setup_work_dir().await?;
-    println!("Updating the CLI");
     let cardano_path = check_env("CARDANO_DIR")?;
     check_repo(CARDANO_URL, &cardano_path).await?;
     let cmd = format!("cd {cardano_path} && git checkout main && git pull && cargo install --path src/bin");
