+++ src/lib/src/configs/mod.rs
@@ -1,18 +1,6 @@
 pub mod globals;
 pub use globals::*;
-pub mod cabal;
-pub use cabal::*;
-pub mod ghc;
-pub use ghc::*;
-pub mod ghcup;
-pub use ghcup::*;
-pub mod libsodium;
-pub use libsodium::*;
-pub mod secp256k1;
-pub use secp256k1::*;
-pub mod setup;
-pub use setup::*;
+pub mod logger;
+pub use logger::*;
 pub mod settings;
 pub use settings::*;
-pub mod system;
-pub use system::*;
