+++ src/lib/src/lib.rs
@@ -1,7 +1,11 @@
 //! # Lib for cardano
 //!
 //! `lib` is a collection of utilities to make performing certain operations more convenient
+pub mod cli;
+pub use cli::*;
 pub mod utils;
 pub use utils::*;
+pub mod packages;
+pub use packages::*;
 pub mod configs;
 pub use configs::*;
