+++ src/lib/src/packages/mod.rs
@@ -1,0 +1,16 @@
+pub mod platform;
+pub use platform::*;
+pub mod cabal;
+pub use cabal::*;
+pub mod ghc;
+pub use ghc::*;
+pub mod ghcup;
+pub use ghcup::*;
+pub mod libsodium;
+pub use libsodium::*;
+pub mod secp256k1;
+pub use secp256k1::*;
+pub mod shell;
+pub use shell::*;
+pub mod system;
+pub use system::*;
