+++ src/lib/src/packages/shell.rs
