+++ src/lib/src/utils/mod.rs
@@ -2,23 +2,7 @@ pub mod env;
 pub use env::*;
 pub mod fs;
 pub use fs::*;
-pub mod process;
-pub use process::*;
-pub mod user;
-pub use user::*;
-pub mod dialog;
-pub use dialog::*;
-pub mod shell;
-pub use shell::*;
-pub mod packages;
-pub use packages::*;
 pub mod git;
 pub use git::*;
-pub mod update_cli;
-pub use update_cli::*;
-pub mod uninstall;
-pub use uninstall::*;
-pub mod test;
-pub use test::*;
-pub mod logger;
-pub use logger::*;
+pub mod process;
+pub use process::*;
